Divide battery percentage by 100 for the value to be properly displayed in Dawarich.

+++ lib/data_contracts/data_transfer_objects/api/v1/overland/batches/request/batch_point_properties_dto.dart
@@ -16,7 +16,7 @@ class BatchPointPropertiesDto {
   final String deviceId;
   final String wifi;
   final String batteryState;
-  final int batteryLevel;
+  final double batteryLevel;
 
   BatchPointPropertiesDto({
     required this.timestamp,
