+++ lib/ui/models/api/v1/overland/batches/request/batch_point_properties_viewmodel.dart
@@ -15,7 +15,7 @@ class BatchPointPropertiesViewModel {
   final String deviceId;
   final String wifi;
   final String batteryState;
-  final int batteryLevel;
+  final double batteryLevel;
 
   BatchPointPropertiesViewModel({
     required this.timestamp,
